+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Introduce CUBE CUT. Press C in edit mode to enter cube cut mode. See Readme for more details.
 - Fix grids not displaying in orthographic view in some cases
+- On Level Design workspace creation, set appropriate blender file settings
 
 ## 1.0.3